scripts/requirements.txt
orange3==3.26
pandas==1.2.2
numpy==1.20.1
matplotlib==3.3.4